requirements.txt
appdirs==1.4.4
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==3.0.4
distlib==0.3.1
distro==1.5.0
filelock==3.0.12
idna==2.10
numpy==1.19.3
pandas==1.1.5
python-dateutil==2.8.1
pytz==2020.4
requests==2.25.0
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
tabula-py==2.2.0
urllib3==1.26.2
virtualenv==20.2.2